Restructure abstract method handling in plugin
